Includes changes to allow the seperate compilation of ntp_data_structures.c
ntp_scanner.c and ntp_parser.yAlso includes other minor bugfixes.

+++ ntpd/cmd_args.c
@@ -16,7 +16,7 @@
  */
 extern char const *progname;
 extern const char *specific_interface;
-extern int default_ai_family;
+extern short default_ai_family;
 
 #ifdef HAVE_NETINFO
 extern int	check_netinfo;
